+++ R/use_function.R
@@ -14,17 +14,18 @@
 #' @param pkg package with a CWB indexed corpus to use (defaults to NULL)
 #' @param lib.loc a character vector with path names of \code{R} libraries
 #' @param dir a registry directory, defaults to \code{getOption("polmineR.defaultRegistry")}
+#' @param verbose logical, whether to output status messages
 #' @return the function returns invisibly the registry that was previously set
 #' @export use
 #' @rdname use
 #' @name use
 #' @seealso the worker to reset the registry is \code{resetRegistry}
-use <- function(pkg = NULL, lib.loc = .libPaths(), dir = getOption("polmineR.defaultRegistry")){
+use <- function(pkg = NULL, lib.loc = .libPaths(), dir = getOption("polmineR.defaultRegistry"), verbose = TRUE){
   if (!is.null(pkg)){
     if (pkg %in% unname(installed.packages(lib.loc = lib.loc)[,"Package"])) {
       registryDir <- system.file("extdata", "cwb", "registry", package = pkg, lib.loc = lib.loc)
       if (dir.exists(registryDir)){
-        previousRegistry <- resetRegistry(registryDir)
+        previousRegistry <- resetRegistry(registryDir, verbose = verbose)
         options("polmineR.templates" = list())
         setTemplate()
       } else {
@@ -34,7 +35,7 @@ use <- function(pkg = NULL, lib.loc = .libPaths(), dir = getOption("polmineR.def
     
   } else {
     if (dir.exists(dir)){
-      previousRegistry <- resetRegistry(dir)
+      previousRegistry <- resetRegistry(dir, verbose = verbose)
       options("polmineR.templates" = list())
       setTemplate()
     } else {